requirements.txt
Flask==3.0.0
openai==0.28.0
python-dotenv==1.0.0
requests==2.31.0
spotipy==2.23.0
urllib3==2.1.0
gunicorn==21.2.0

